backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
SQLAlchemy==2.0.36
python-multipart==0.0.9
python-jose==3.3.0
bcrypt==4.2.0
passlib==1.7.4
aiofiles==24.1.0
python-dotenv==1.0.1
Pillow==10.4.0
Markdown==3.6